# [RF] Faster name pointer lookup for `RooSTLRefCountList`