--- a/roofit/roofitcore/inc/RooNameReg.h
+++ b/roofit/roofitcore/inc/RooNameReg.h
@@ -32,6 +32,7 @@ public:
   static const TNamed* ptr(const char* stringPtr) ;
   static const char* str(const TNamed* ptr) ;
   static const TNamed* known(const char* stringPtr) ;
+  static const std::size_t& renameCounter() ;
 
   enum {
     kRenamedArg = BIT(19)    // TNamed flag to indicate that some RooAbsArg has been renamed (flag set in new name)
@@ -42,7 +43,11 @@ protected:
 //  RooNameReg(Int_t hashSize = 31) ;
   RooNameReg(const RooNameReg& other) = delete;
 
+  friend class RooAbsArg;
+  static void incrementRenameCounter() ;
+
   std::unordered_map<std::string,std::unique_ptr<TNamed>> _map;
+  std::size_t _renameCounter = 0;
 
 //  ClassDef(RooNameReg,1) // String name registry
 };