The `RooSTLRefCountList` is used in the `RooAbsArg` client server    
interface, which is why speeding up lookup operations is very important    
for large models.    
    
With this commit:    
    
The `RooSTLRefCountList` is used in the `RooAbsArg` client server
interface, which is why speeding up lookup operations is very important
for large models.

With this commit:

* A `RooSTLRefCountList` has an additinal member that tracks the
  ordering of the elements by `RooAbsArg::namePtr()` address

* Subsequently, looking up elements by name pointer is much faster

* Care was taken to handle the case where elements already in the
  collection get renamed (see inline documentation and unit test)

* The `RooSTLRefCountList` can notice if it was read from IO and
  determine the ordering before doing any lookups

As far as I can tell, this change has now drawbacks other than some
increased memory usage. The order of the elements is still the same, so
no user code should be affected.
    
As far as I can tell, this change has now drawbacks other than some    
increased memory usage. The order of the elements is still the same, so     
no user code should be affected.

Benchmark before this commit ([profile](https://rembserj.web.cern.ch/rembserj/cgi-bin/igprof-navigator/VHbb-STXS_mu__createNLL)):
```
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
VHbb-STXS_mu__createNLL/0/iterations:1       14.1 s          14.0 s             1

```
Benchmark after this commit ([profile](https://rembserj.web.cern.ch/rembserj/cgi-bin/igprof-navigator/VHbb-STXS_mu__createNLL_2)):
```
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
VHbb-STXS_mu__createNLL/0/iterations:1       5.13 s          5.13 s             1

```